umm..refactor the whole app...
